refact(trees): Migrate tree files to src/trees/ and improve Makefile cross-compatibility
