+++ src/data_structures/data_structures_demo.c
@@ -1,4 +1,5 @@
 #include "data_structures.h"
+#include "trees.h"
 #include "safe_input.h"
 #include <stdio.h>
